Refactor userHasElevatedRole method to check for ADMINISTRATOR flag
