+++ src/bot/commands/AvatarCommand.ts
@@ -1,9 +1,9 @@
-import { ClientUser, Message, Permissions } from 'discord.js';
+import { ClientUser, Message } from 'discord.js';
 
 import Config from '@config/Config';
 import localize from '@util/i18n/localize';
 import UserCommand from './base/UserCommand';
-import userHasElevatedRole from './helpers/checkElevatedRights';
+import userHasElevatedRole from './helpers/userHasElevatedRole';
 
 export default class AvatarCommand implements UserCommand {
   public readonly TRIGGERS = ['avatar'];
@@ -24,10 +24,8 @@ export default class AvatarCommand implements UserCommand {
   public run(message: Message, params: string[]) {
     if (!message.member) return;
 
-    const allowedToRunCommand = userHasElevatedRole(message.member.roles.cache);
-    if (!message.member.hasPermission(Permissions.FLAGS.ADMINISTRATOR!) && !allowedToRunCommand) {
-      return;
-    }
+    const allowedToRunCommand = userHasElevatedRole(message.member);
+    if (!allowedToRunCommand) return;
 
     if (params.length === this.NUMBER_OF_PARAMETERS && params[0] === 'remove') {
       this.user.setAvatar('');
