+++ src/bot/commands/ConfigCommand.ts
@@ -1,9 +1,9 @@
-import { ClientUser, Message, Permissions } from 'discord.js';
+import { ClientUser, Message } from 'discord.js';
 
 import Config from '@config/Config';
 import localize from '@util/i18n/localize';
 import Command from './base/Command';
-import userHasElevatedRole from './helpers/checkElevatedRights';
+import userHasElevatedRole from './helpers/userHasElevatedRole';
 
 export default class ConfigCommand implements Command {
   public readonly TRIGGERS = ['config', 'set'];
@@ -24,10 +24,8 @@ export default class ConfigCommand implements Command {
   public run(message: Message, params: string[]) {
     if (!message.member) return;
 
-    const allowedToRunCommand = userHasElevatedRole(message.member.roles.cache);
-    if (!message.member.hasPermission(Permissions.FLAGS.ADMINISTRATOR!) && !allowedToRunCommand) {
-      return;
-    }
+    const allowedToRunCommand = userHasElevatedRole(message.member);
+    if (!allowedToRunCommand) return;
 
     if (params.length < this.NUMBER_OF_PARAMETERS) {
       message.channel.send(this.USAGE);
