+++ src/bot/commands/RenameCommand.ts
@@ -1,13 +1,13 @@
 import fs from 'fs';
 
-import { Message, Permissions } from 'discord.js';
+import { Message } from 'discord.js';
 
 import * as soundsDb from '@util/db/Sounds';
 import localize from '@util/i18n/localize';
 import { getExtensionForSound, getSounds } from '@util/SoundUtil';
 import Config from '@config/Config';
 import Command from './base/Command';
-import userHasElevatedRole from './helpers/checkElevatedRights';
+import userHasElevatedRole from './helpers/userHasElevatedRole';
 
 export default class RenameCommand implements Command {
   public readonly TRIGGERS = ['rename'];
@@ -23,10 +23,8 @@ export default class RenameCommand implements Command {
   public run(message: Message, params: string[]) {
     if (!message.member) return;
 
-    const allowedToRunCommand = userHasElevatedRole(message.member.roles.cache);
-    if (!message.member.hasPermission(Permissions.FLAGS.ADMINISTRATOR!) && !allowedToRunCommand) {
-      return;
-    }
+    const allowedToRunCommand = userHasElevatedRole(message.member);
+    if (!allowedToRunCommand) return;
 
     if (params.length !== this.NUMBER_OF_PARAMETERS) {
       message.channel.send(this.USAGE);
